Update MSMLib.py
Added some code to include a mask argument that allows separation of cis and trans conformers of CheY, based on the proline dihedral angle. There are additional small edits in msm_analysis and CalculateImpliedTimescales.py. I set it up such that 'mask' is completely optional, so if not explicitly specified, all code should run as usual.