Fixing non-convergence issue in calculate_fluxes in tpt.py
In calculate_fluxes, the state populations are calculated by only requesting one eigenvalue (line 450), but I found that this could lead to non-convergence issues, so it would probably be a good idea to request a larger number of eigenvalues to ensure the equilibrium population can be obtained (i.e. n >= 5).